requirements.txt
accelerate==0.29.3
advertorch==0.2.4
advex_uar==0.0.6.dev0
autoattack==0.1
bitsandbytes==0.45.5
colorama==0.4.6
datasets==2.19.1
diffusers==0.30.0.dev0
lpips==0.1.4
matplotlib==3.6.3
numpy==1.24.4
opencv_python==4.1.2.30
opencv_python_headless==4.9.0.80
packaging==25.0
pandas==1.2.5
peft==0.10.0
Pillow==9.5.0
Pillow==11.2.1
safetensors==0.5.3
timm>=1.0.0
torch==2.4.0
torcheval==0.0.7
torchmetrics==0.6.0
torchvision==0.19.0
tqdm==4.56.2
transformers==4.40.1
wandb>=0.17.0
xformers==0.0.27.post2
